// src/shared/api/models/postApiExpenseBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiExpenseBodyExpenseCategoryId } from './postApiExpenseBodyExpenseCategoryId';

export type PostApiExpenseBody = {
    employeeId: number;
    expenseCategoryId?: PostApiExpenseBodyExpenseCategoryId;
    /** @minimum 0.01 */
    amount: number;
    description?: string;
};